# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plant Doctor API",
        "description": "API for plant disease diagnosis and health tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8005",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/diseases": {
            "get": {
                "tags": [
                    "Diseases"
                ],
                "summary": "List diseases",
                "description": "GET /api/diseases\nList diseases with search and filter",
                "operationId": "1d8c114c95b03eaae97cd6a1e07fa9e4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name, scientific name or description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diseases listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Disease"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/diseases/{disease}": {
            "get": {
                "tags": [
                    "Diseases"
                ],
                "summary": "Get disease details",
                "description": "GET /api/diseases/{disease}\nGet disease details",
                "operationId": "c45212fa9aa10792a6c79a91289f96d1",
                "parameters": [
                    {
                        "name": "disease",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disease retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Disease"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Disease": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Oídio"
                    },
                    "scientific_name": {
                        "type": "string",
                        "example": "Erysiphe cichoracearum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "symptoms": {
                        "type": "string",
                        "nullable": true
                    },
                    "treatment": {
                        "type": "string",
                        "nullable": true
                    },
                    "prevention": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "fungal",
                        "nullable": true
                    },
                    "severity": {
                        "type": "string",
                        "example": "medium"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "perenual_id": {
                        "description": "Perenual API ID",
                        "type": "integer",
                        "nullable": true
                    },
                    "last_synced_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}